+++ src/SmartHeater.Shared/Models/HeaterListModel.cs
@@ -5,4 +5,4 @@ namespace SmartHeater.Shared.Models;
 public record HeaterListModel(string IpAddress,
                               string Name,
                               HeaterTypes HeaterType,
-                              double ReferenceTemperature);
+                              float ReferenceTemperature);
